Tetris game
Double click the Tetris3D.exe to run the game. I was too lazy to implement a help page so here it goes - use the WASD keys to move the piece, use the arrow keys to rotate the keys (relative to the camera), click the shift key to drop the pice by one level or the tab key to drop the piece as much as possible, press Q and E to rotate the camera around the platform and hold the spacebar to get an ariel view (will revert to normal view when the space key is released).